+++ wp-content/plugins/woocommerce/includes/log-handlers/class-wc-log-handler-file.php
@@ -5,6 +5,8 @@
  * @package WooCommerce\Log Handlers
  */
 
+use Automattic\Jetpack\Constants;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
@@ -14,7 +16,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @class          WC_Log_Handler_File
  * @version        1.0.0
- * @package        WooCommerce/Classes/Log_Handlers
+ * @package        WooCommerce\Classes\Log_Handlers
  */
 class WC_Log_Handler_File extends WC_Log_Handler {
 
@@ -144,10 +146,12 @@ class WC_Log_Handler_File extends WC_Log_Handler {
 		if ( $file ) {
 			if ( ! file_exists( $file ) ) {
 				$temphandle = @fopen( $file, 'w+' ); // @codingStandardsIgnoreLine.
-				@fclose( $temphandle ); // @codingStandardsIgnoreLine.
+				if ( is_resource( $temphandle ) ) {
+					@fclose( $temphandle ); // @codingStandardsIgnoreLine.
 
-				if ( defined( 'FS_CHMOD_FILE' ) ) {
-					@chmod( $file, FS_CHMOD_FILE ); // @codingStandardsIgnoreLine.
+					if ( Constants::is_defined( 'FS_CHMOD_FILE' ) ) {
+						@chmod( $file, FS_CHMOD_FILE ); // @codingStandardsIgnoreLine.
+					}
 				}
 			}
 
@@ -248,16 +252,17 @@ class WC_Log_Handler_File extends WC_Log_Handler {
 	 */
 	public function remove( $handle ) {
 		$removed = false;
-		$file    = self::get_log_file_path( $handle );
+		$logs    = $this->get_log_files();
+		$handle  = sanitize_title( $handle );
 
-		if ( $file ) {
-			if ( is_file( $file ) && is_writable( $file ) ) { // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_is_writable
-				$this->close( $handle ); // Close first to be certain no processes keep it alive after it is unlinked.
+		if ( isset( $logs[ $handle ] ) && $logs[ $handle ] ) {
+			$file = realpath( trailingslashit( WC_LOG_DIR ) . $logs[ $handle ] );
+			if ( 0 === stripos( $file, realpath( trailingslashit( WC_LOG_DIR ) ) ) && is_file( $file ) && is_writable( $file ) ) { // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_is_writable
+				$this->close( $file ); // Close first to be certain no processes keep it alive after it is unlinked.
 				$removed = unlink( $file ); // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_unlink
 			}
 			do_action( 'woocommerce_log_remove', $handle, $removed );
 		}
-
 		return $removed;
 	}
 
@@ -363,7 +368,7 @@ class WC_Log_Handler_File extends WC_Log_Handler {
 	 */
 	public static function get_log_file_name( $handle ) {
 		if ( function_exists( 'wp_hash' ) ) {
-			$date_suffix = date( 'Y-m-d', current_time( 'timestamp', true ) );
+			$date_suffix = date( 'Y-m-d', time() );
 			$hash_suffix = wp_hash( $handle );
 			return sanitize_file_name( implode( '-', array( $handle, $date_suffix, $hash_suffix ) ) . '.log' );
 		} else {
